chore: refactor directories to be under ecsact dir
