+++ ecsact/cli/detail/executable_path/unix_executable_path.cc
