+++ ecsact/cli/detail/executable_path/windows_executable_path.cc
