--- ecsact/cli/detail/report_proc_output.hh
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <iostream>
-#include <boost/process.hpp>
-
-namespace ecsact::cli::detail {
-
-  
-}
